requirements.txt
fastapi
uvicorn
dashscope
tavily-python
pydantic
python-dotenv
requests
beautifulsoup4
python-multipart
opencv-python-headless
redis
rq
sqlmodel
httpx
